% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{as_nrs.DISTINCT}
\alias{as_nrs.DISTINCT}
\title{Convert a DISTINCT object to a nrs (NhoodRefSim sample) object}
\usage{
\method{as_nrs}{DISTINCT}(
  obj,
  ltb = obj$hyperparams$tier2$ltb,
  primary_ref_samples = eval(formals(gamma_tit_error_estimates)$primary_ref_samples),
 
    exclusion_affinity_bms = eval(formals(gamma_tit_error_estimates)$exclusion_affinity_bms)
)
}
\description{
Convert a DISTINCT object to a nrs (NhoodRefSim sample) object
}
