% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_test.R
\name{discerner_test}
\alias{discerner_test}
\title{Test for differential state between two groups of samples, based on scRNA-seq data.}
\usage{
discerner_test(x, name_assays_expression = "counts",
  name_cluster = "cluster_id", name_sample = "sample_id",
  name_group = "group_id",
  name_metadata_experiment_info = "experiment_info", perm_cells = TRUE,
  design = NULL, logarithm = FALSE, P = 10^3, N_breaks = 10,
  min_non_zero_cells_per_group = 20, nCores = 1)
}
\arguments{
\item{x}{a \code{\linkS4class{SingleCellExperiment}} object.}

\item{name_assays_expression}{a character ("counts" by default), 
indicating the name of the assays(x) element which stores the expression data (i.e., assays(x)$name_assays_expression).}

\item{perm_cells}{a boolean, indicating whether to permute cells (if TRUE, by default) or samples (if FALSE).}

\item{logarithm}{a boolean, indicating whether to use raw counts (if FALSE, by default) or log(counts + 1) (if TRUE).}

\item{P}{the number of permutations to use.}

\item{N_breaks}{the number of breaks at which to evaluate the comulative density function.}

\item{min_non_zero_cells_per_group}{the minimum number of non-zero cells in each group for a gene to be evaluated.}

\item{name_colData_cluster}{a character ("cluster_id" by default), 
indicating the name of the colData(x) element which stores the cluster id of each cell (i.e., colData(x)$name_colData_cluster).}

\item{name_colData_sample}{a character ("sample_id" by default), 
indicating the name of the colData(x) element which stores the sample id of each cell (i.e., colData(x)$name_colData_sample).}

\item{name_colData_group}{a character ("group_id" by default), 
indicating the name of the colData(x) element which stores the group id of each cell (i.e., colData(x)$name_colData_group).}
}
\value{
A \code{\linkS4class{data.frame}} object.
}
\description{
\code{discerner_test} tests for differential state between two groups of samples, based on scRNA-seq data.
}
\examples{
# discerner_test(x)

}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
