\name{distinct-package}
\alias{distinct-package}
\alias{distinct}
\docType{package}
\title{
\packageTitle{distinct}
}
\description{
\packageDescription{distinct}

Questions relative to distinct should be either written to the \href{https://support.bioconductor.org}{Bioconductor support site}, tagging the question with 'distinct', or reported as a new issue at \href{https://github.com/SimoneTiberi/distinct/issues}{BugReports}.

To access the vignettes, type: \code{browseVignettes("distinct")}.
}
\author{
\packageAuthor{distinct}

Maintainer: \packageMaintainer{distinct}
}
\keyword{
  differential distribution,
  differential state,
  differential expression,
  hierarchical permutation tests
}