% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{DISTINCT}
\alias{DISTINCT}
\title{Assess DISTINCT results, class instantiator}
\usage{
DISTINCT(
  bm_res = NULL,
  bm_res_sum = NULL,
  rank_fun = rank_funs[["TNFa_duration"]],
  filtering_fun = filtering_funs[[1]],
  lookup_hyperparams_f = lookup_tier1,
  settings_grid = NH_t_grid,
  tier1_rank = 1L,
  plot_id = NULL
)
}
\description{
This class copies all related objects into one 'basket'/class so
that we don't have to worry about mismatched objects down the line
}
