% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NhoodRefSim.R
\name{exclude_ref_sample_cor}
\alias{exclude_ref_sample_cor}
\title{Simplify a collection of reference samples using Euclidean distance}
\usage{
exclude_ref_sample_cor(
  M,
  sa,
  sample_name_regex = NULL,
  sample_names = NULL,
  max_tol_aff = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{M}{Matrix of sample coordinates in latent or raw space. Row
names must be sample identifiers}

\item{sa}{Sample annotation. At least one the columns must contain
the rownames of M.}

\item{sample_name_regex}{Regex to use to identify 'anchor' samples}

\item{sample_names}{Vector of anchor sample names. Either this
argument of \code{sample_name_regex} mus be NULL}

\item{max_tol_aff}{Maximum tolerated affinity for a query sample
and any of the anchor samples in order for the query sample not to
be thrown out.}

\item{verbose}{Whether to print intermediate information}
}
\value{
A list with the cleaned up matrix M and sample annotation
sa
}
\description{
Reduce the number of reference samples by specifying a set of
'anchor' samples and assessing their Euclidean distance to the
remaining 'query' samples. Any of the query samples sufficiently
far away from the anchor samples are kept.
}
