% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_nlp_features}
\alias{extract_nlp_features}
\title{Extract NLP Features from CT Reports}
\usage{
extract_nlp_features(df)
}
\arguments{
\item{df}{A data frame with CT report text and necessary identifiers.}
}
\value{
A data frame with the original data and new NLP feature columns.
}
\description{
This function processes CT report text to extract NLP features based on key phrases.
It segments the text into Clinical History, Findings, and Impression sections and
calculates the frequency of specific keyphrases as NLP features.
}
\examples{
#  Assuming 'data' is a data frame with CT report text and
# 'NLPList' and 'NLPFeatureNames' are defined
data_with_nlp_features <- extract_nlp_features(data, NLPList, NLPFeatureNames)

}
