% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_match_col}
\alias{find_match_col}
\title{Search for values v in all columns of dtf}
\usage{
find_match_col(v, dtf)
}
\arguments{
\item{v}{A character vector}

\item{dtf}{A data.frame}
}
\value{
The name of the column in dtf that contains the most
matches with the entries in v
}
\description{
Search for values v in all columns of dtf
}
