% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query2ref.R
\name{gamma_tit_error_estimates}
\alias{gamma_tit_error_estimates}
\title{Run agg_Nhood_error for all dtfs in a gamma titration object}
\usage{
gamma_tit_error_estimates(
  gamma_tit,
  ref_so,
  q_sc_so,
  sce,
  ref_reconstruction = NULL,
  primary_ref_samples = list(c("Unstimulated in vitro - 2h",
    "Unstimulated in vitro - 6h", "Unstimulated in vitro - 12h",
    "Unstimulated in vitro - 24h", "0.1 ng/ml TNFa - 2h", "0.1 ng/ml TNFa - 6h",
    "0.1 ng/ml TNFa - 12h", "0.1 ng/ml TNFa - 24h", "1 ng/ml TNFa - 2h",
    "1 ng/ml TNFa - 6h", "1 ng/ml TNFa - 12h", "1 ng/ml TNFa - 24h",
    "10 ng/ml TNFa - 2h", "10 ng/ml TNFa - 6h", "10 ng/ml TNFa - 12h",
    "10 ng/ml TNFa - 24h", "1 ng/ml IFNy - 2h", "1 ng/ml IFNy - 6h",
    "1 ng/ml IFNy - 12h", "1 ng/ml IFNy - 24h", "10 ng/ml IFNy - 2h",     
    "10 ng/ml IFNy - 6h", "10 ng/ml IFNy - 12h", "10 ng/ml IFNy - 24h",
    "100 ng/ml IFNy - 2h", "100 ng/ml IFNy - 6h", "100 ng/ml IFNy - 12h",
    "100 ng/ml IFNy - 24h")),
  exclusion_affinity_bms = c(compile_all_pairs(c("Unstimulated in vitro - 2h",
    "Unstimulated in vitro - 6h", "Unstimulated in vitro - 12h",
    "Unstimulated in vitro - 24h")), NULL),
  min_Nhood_sizes = as.integer(c(0, 5, 25, 50, 100)),
  do_debug = FALSE,
  ref_weights = "none",
  verbose = TRUE
)
}
\arguments{
\item{gamma_tit}{Gamma titration, 1 row per value of gamma}
}
\description{
Run agg_Nhood_error for all dtfs in a gamma titration object
}
