% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_euclidean}
\alias{matrix_euclidean}
\title{Compute the Euclidean distances between  the columns of two
matrices}
\usage{
matrix_euclidean(X, Y)
}
\description{
Compute the Euclidean distances between  the columns of two
matrices
}
