% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NhoodRefSim.R
\name{NhoodRefSim}
\alias{NhoodRefSim}
\title{Init an Nhood to reference similarity object}
\value{
an NhoodRefSim object
}
\description{
Init an Nhood to reference similarity object

Init an Nhood to reference similarity object
}
\examples{

## ------------------------------------------------
## Method `NhoodRefSim$new`
## ------------------------------------------------

NhoodRefSim$initialize()

Subset the reference sample annotation with a specific set of
samples


}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{NhoodRefSim$print()}}
\item \href{#method-new}{\code{NhoodRefSim$new()}}
\item \href{#method-subset_ref_sa}{\code{NhoodRefSim$subset_ref_sa()}}
\item \href{#method-compute_distM}{\code{NhoodRefSim$compute_distM()}}
\item \href{#method-compute_affM}{\code{NhoodRefSim$compute_affM()}}
\item \href{#method-agg_Nhood_error}{\code{NhoodRefSim$agg_Nhood_error()}}
\item \href{#method-compute_Nhood_stats}{\code{NhoodRefSim$compute_Nhood_stats()}}
\item \href{#method-overlay_cluster_cuts}{\code{NhoodRefSim$overlay_cluster_cuts()}}
\item \href{#method-overlay_cluster_labels}{\code{NhoodRefSim$overlay_cluster_labels()}}
\item \href{#method-compute_cluster_enrichment}{\code{NhoodRefSim$compute_cluster_enrichment()}}
\item \href{#method-compute_rep_ref_distM}{\code{NhoodRefSim$compute_rep_ref_distM()}}
\item \href{#method-plot_ref_affM_UMAP}{\code{NhoodRefSim$plot_ref_affM_UMAP()}}
\item \href{#method-compute_UMAP_coords}{\code{NhoodRefSim$compute_UMAP_coords()}}
\item \href{#method-simplify_affM}{\code{NhoodRefSim$simplify_affM()}}
\item \href{#method-plot_Nhood_ref_affinity_graph}{\code{NhoodRefSim$plot_Nhood_ref_affinity_graph()}}
\item \href{#method-compute_ref_tree}{\code{NhoodRefSim$compute_ref_tree()}}
\item \href{#method-clone}{\code{NhoodRefSim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$new(
  dtf,
  sce,
  query,
  ref_sa,
  ref_experiment = ref_sa$experiment[1],
  primary_ref_samples = list(),
  exclusion_affinity_bms = list(),
  ref_reconstruction = NULL,
  N_PV = NULL,
  MEV_q = NULL,
  ltb = NULL,
  cluster_h_q = NULL,
  ref_CF_clustering = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dtf}}{Dataframe}

\item{\code{query}}{Name of the query experiment}

\item{\code{ref_sa}}{Meta data of the reference experiment}

\item{\code{ref_experiment}}{Name of the reference experiment
samples by their uniqueness. Will be deduced from @ref_sa if
absent.}

\item{\code{ref_CF_clustering}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{NhoodRefSim$initialize()

Subset the reference sample annotation with a specific set of
samples


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset_ref_sa"></a>}}
\if{latex}{\out{\hypertarget{method-subset_ref_sa}{}}}
\subsection{Method \code{subset_ref_sa()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$subset_ref_sa(sample_names = colnames(self$distM))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_distM"></a>}}
\if{latex}{\out{\hypertarget{method-compute_distM}{}}}
\subsection{Method \code{compute_distM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_distM()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_affM"></a>}}
\if{latex}{\out{\hypertarget{method-compute_affM}{}}}
\subsection{Method \code{compute_affM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_affM(distM = self$distM, ltb = self$ltb, min_sum_prob = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-agg_Nhood_error"></a>}}
\if{latex}{\out{\hypertarget{method-agg_Nhood_error}{}}}
\subsection{Method \code{agg_Nhood_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$agg_Nhood_error(
  ltb = self$ltb,
  reg_vars = c("duration", "ifn_conc", "tnf_conc"),
  by_var = "condition_name",
  min_Nhood_size = NULL,
  scale_by_max_aff = FALSE,
  ref_weights = "none",
  include_bs_pres = FALSE,
  verbose = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_Nhood_stats"></a>}}
\if{latex}{\out{\hypertarget{method-compute_Nhood_stats}{}}}
\subsection{Method \code{compute_Nhood_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_Nhood_stats(min_Nhood_size = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-overlay_cluster_cuts"></a>}}
\if{latex}{\out{\hypertarget{method-overlay_cluster_cuts}{}}}
\subsection{Method \code{overlay_cluster_cuts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$overlay_cluster_cuts(heatmap_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-overlay_cluster_labels"></a>}}
\if{latex}{\out{\hypertarget{method-overlay_cluster_labels}{}}}
\subsection{Method \code{overlay_cluster_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$overlay_cluster_labels(x_loc = 0.9, y_loc = 0.1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_cluster_enrichment"></a>}}
\if{latex}{\out{\hypertarget{method-compute_cluster_enrichment}{}}}
\subsection{Method \code{compute_cluster_enrichment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_cluster_enrichment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_rep_ref_distM"></a>}}
\if{latex}{\out{\hypertarget{method-compute_rep_ref_distM}{}}}
\subsection{Method \code{compute_rep_ref_distM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_rep_ref_distM(ltb = 20)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_ref_affM_UMAP"></a>}}
\if{latex}{\out{\hypertarget{method-plot_ref_affM_UMAP}{}}}
\subsection{Method \code{plot_ref_affM_UMAP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$plot_ref_affM_UMAP()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_UMAP_coords"></a>}}
\if{latex}{\out{\hypertarget{method-compute_UMAP_coords}{}}}
\subsection{Method \code{compute_UMAP_coords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_UMAP_coords(cap = 1000)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simplify_affM"></a>}}
\if{latex}{\out{\hypertarget{method-simplify_affM}{}}}
\subsection{Method \code{simplify_affM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$simplify_affM(max_allowed_size = 0.01)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_Nhood_ref_affinity_graph"></a>}}
\if{latex}{\out{\hypertarget{method-plot_Nhood_ref_affinity_graph}{}}}
\subsection{Method \code{plot_Nhood_ref_affinity_graph()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$plot_Nhood_ref_affinity_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_ref_tree"></a>}}
\if{latex}{\out{\hypertarget{method-compute_ref_tree}{}}}
\subsection{Method \code{compute_ref_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$compute_ref_tree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NhoodRefSim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
