% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{plot.DISTINCT}
\alias{plot.DISTINCT}
\title{Make all DISTINCT plots}
\usage{
\method{plot}{DISTINCT}(
  obj,
  ltb = obj$hyperparams$tier2$ltb,
  out_dir = get_out_dir(obj),
  plot_id = get_plot_id(obj)
)
}
\description{
Make all DISTINCT plots
}
