% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dim_reduc}
\alias{plot_dim_reduc}
\title{Plot a dimension reduction of integrated data}
\usage{
plot_dim_reduc(
  dtf = umap_emb,
  colour_var = "stim_group",
  use_stim_group_cols = TRUE,
  shape_var = "duration",
  coord_regex = "UMAP",
  legend_ncol = 1,
  zoom_range = 1 == 1,
  plot_range = NULL,
  zoom_name = NULL,
  filter_samples = 1 == 1,
  ordering_code = NULL,
  filter_name = NULL,
  ref_experiments = c("4910", "5029", "6434", "6623", "5029-6434"),
  point_alpha = 0.5,
  plot_ref_background = T,
  width = 17.4,
  height = 20,
  annotation_code = NULL,
  fn_app = "",
  print_to_file = F
)
}
\arguments{
\item{dtf}{Data.frame of latent variables + meta.data + UMAP
coordinates (one sample per row)}
}
\description{
Plot a dimension reduction of integrated data
}
