% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{print.DISTINCT_bm_summary}
\alias{print.DISTINCT_bm_summary}
\title{Print an object of the DISTINCT_bm_summary class}
\usage{
\method{print}{DISTINCT_bm_summary}(obj)
}
\arguments{
\item{obj}{A DISTINCT_bm_summary object}
}
\description{
Print an object of the DISTINCT_bm_summary class
}
