% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{print.rank_fun}
\alias{print.rank_fun}
\title{Print a human-friendly version of a ranking function}
\usage{
\method{print}{rank_fun}(obj)
}
\arguments{
\item{obj}{List of parameters}
}
\description{
Print a human-friendly version of a ranking function
}
