% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read and Validate Data for CT Indication Prediction}
\usage{
read_data(input_data)
}
\arguments{
\item{input_data}{A file path to a CSV/Excel file or a data frame.}
}
\value{
A data frame with validated and correctly typed columns.
}
\description{
This function reads data from a CSV or Excel file or an existing data frame,
checks for the presence of required columns, and validates their data types.
Optional columns are checked if they exist. The function also specifies a
date format for date columns and converts columns to the correct data type.
}
\details{
Compulsory columns:
\itemize{
  \item \code{patient_id}: Unique identifier for the patient (required, character or numeric).

  \item \code{diagnosis_date}: Date of the initial diagnosis (required, expected in "YYYY-MM-DD" format).

  \item \code{ct_date}: Date of the CT scan (required, expected in "YYYY-MM-DD" format).

  \item \code{provider_type}: Type of provider ordering the CT scan (required, character),
        examples include Internal Medicine, Emergency Medicine, Radiation Oncology, Medical Oncology, etc.

  \item \code{report}: Text of the CT scan report (required, character).

  \item \code{symptom_diagnosis}: Binary indicator of whether the patient was diagnosed with symptoms
        like fever, cough, chest pain, shortness of breath within 6 months prior to the current CT
        (required, 1 for presence, 0 for absence, or numeric counts).

  \item \code{lungdisease_diagnosis}: Binary indicator of whether the patient was diagnosed with lung
        diseases like pleural effusion, pneumonia, hemoptysis, dyspnea, emphysema, COPD within 6 months
        prior to the current CT
        (required, 1 for presence, 0 for absence, or numeric counts).

  \item \code{Xray_count}: Count of X-ray scans 6 months prior to the current CT (required, numeric).

}
Optional columns:
\itemize{

  \item \code{symptom_names}: Names of symptoms reported (optional).

  \item \code{lungdisease_names}: Names of lung diseases diagnosed (optional).

}
}
\examples{
# Assuming 'data.xlsx' is in your working directory and has the correct structure
read_data("data.xlsx")
# For an existing data frame named 'dframe'
read_data(dframe)

}
