% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query2ref.R
\name{resim_pmf}
\alias{resim_pmf}
\title{Resimulate from a probability mass function and return metadata
(sample annotation) corresponding to simulated classes}
\usage{
resim_pmf(X, sa, N_resamples = 1000, N_draws = 100)
}
\description{
Resimulate from a probability mass function and return metadata
(sample annotation) corresponding to simulated classes
}
