% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{settings_scatter.DISTINCT_bm_summary}
\alias{settings_scatter.DISTINCT_bm_summary}
\title{Plot a scatter of all statistics for a summary of settings}
\usage{
\method{settings_scatter}{DISTINCT_bm_summary}(bm_res_sum, of = NULL)
}
\arguments{
\item{bm_res_sum}{A DISTINCT_bm_summary object}

\item{of}{Output file, where to store the plot}
}
\description{
Plot a scatter of all statistics for a summary of settings
}
