% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_settings.R
\name{summarize_bm_res}
\alias{summarize_bm_res}
\title{Reduce a large set of benchmarking results to 1 setting per
combination of tier 1 parameters}
\usage{
summarize_bm_res(
  bm_res,
  include_non_hl = FALSE,
  mod_names = FALSE,
  filtering_fun = filtering_funs[[1]],
  rank_fun = rank_funs[[1]]
)
}
\arguments{
\item{bm_res}{A list of benchmarking results}

\item{include_non_hl}{Also include non-highlighted tier 2 setings.
This setting will add he 'hl' column to the function's output.}

\item{filtering_fun}{Function to pre-filter benchmarking results
with}

\item{rank_fun}{Function to rank benchmarking results with}
}
\value{
An ordered tibble of benchmarking result of class
'DISTINCT_bm_summary'
}
\description{
Reduce a large set of benchmarking results to 1 setting per
combination of tier 1 parameters
}
