% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_test.R
\name{distinct_test}
\alias{distinct_test}
\title{Test for differential state between two groups of samples, based on scRNA-seq data.}
\usage{
distinct_test(x, name_assays_expression = "counts",
  name_cluster = "cluster_id", name_sample = "sample_id",
  name_group = "group_id",
  name_metadata_experiment_info = "experiment_info", logarithm = FALSE,
  P = 10^3, N_breaks = 10, min_non_zero_cells_per_group = 20)
}
\arguments{
\item{x}{a \code{\linkS4class{SummarizedExperiment}} object.}

\item{name_assays_expression}{a character ("counts" by default), 
indicating the name of the assays(x) element which stores the expression data (i.e., assays(x)$name_assays_expression).}

\item{name_cluster}{a character ("cluster_id" by default), 
indicating the name of the colData(x) element which stores the cluster id of each cell (i.e., colData(x)$name_colData_cluster).}

\item{name_sample}{a character ("sample_id" by default), 
indicating the name of the colData(x) element which stores the sample id of each cell (i.e., colData(x)$name_colData_sample).}

\item{name_group}{a character ("group_id" by default), 
indicating the name of the colData(x) element which stores the group id of each cell (i.e., colData(x)$name_colData_group).}

\item{name_metadata_experiment_info}{a character ("experiment_info" by default), 
indicating the name of the metadata(x) element which stores the group id of each cell (i.e., metadata(x)$name_metadata_experiment_info).}

\item{logarithm}{a boolean, indicating whether to use raw counts (if FALSE, by default) or log(counts + 1) (if TRUE).}

\item{P}{the number of permutations to use.}

\item{N_breaks}{the number of breaks at which to evaluate the comulative density function.}

\item{min_non_zero_cells_per_group}{the minimum number of non-zero cells in each group for a gene to be evaluated.}
}
\value{
A \code{\linkS4class{data.frame}} object.
}
\description{
\code{distinct_test} tests for differential state between two groups of samples, based on scRNA-seq data.
}
\examples{
library(HDCytoData)
main = Weber_BCR_XL_sim_main_SE()
main

library(diffcyt)
# normalize data:
main <- transformData(main, cofactor = 5)

# select all State markers:
# sel_cols = colData(main)$marker_class == "state"
# select 1 marker only:
sel_cols = colData(main)$marker_name == "pNFkB"

library(SummarizedExperiment)
# create a SummarizedExperiment object with seleceted columns, and inverting row-column structure:
sce <- SummarizedExperiment(
  assays = list(exprs = t(assays(main)$exprs[, sel_cols]) ), 
  colData = rowData(main), 
  rowData = colnames(main)[sel_cols], 
  metadata = list(experiment_info =  metadata(main)$experiment_info ))
  sce
  
# Perform differential analyses, within each cluster of cells, between conditions:
res = distinct_test(
  sce, 
  name_assays_expression = "exprs",
  name_cluster = "population_id",
  name_group = "group_id",
  name_sample = "sample_id",
  P = 10^2, 
  min_non_zero_cells_per_group = 0)

# Visualize results:
head(res)

}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
