% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_test.R
\name{distinct_test}
\alias{distinct_test}
\title{Test for differential state between two groups of samples, based on scRNA-seq data.}
\usage{
distinct_test(x, name_assays_expression = "counts",
  name_cluster = "cluster_id", name_sample = "sample_id",
  name_group = "group_id", logarithm = FALSE, P = 10^3,
  N_breaks = 10, min_non_zero_cells = 20)
}
\arguments{
\item{x}{a \code{\linkS4class{SummarizedExperiment}} or a \code{\linkS4class{SingleCellExperiment}} object.}

\item{name_assays_expression}{a character ("counts" by default), 
indicating the name of the assays(x) element which stores the expression data (i.e., assays(x)$name_assays_expression).}

\item{name_cluster}{a character ("cluster_id" by default), 
indicating the name of the colData(x) element which stores the cluster id of each cell (i.e., colData(x)$name_colData_cluster).}

\item{name_sample}{a character ("sample_id" by default), 
indicating the name of the colData(x) element which stores the sample id of each cell (i.e., colData(x)$name_colData_sample).}

\item{name_group}{a character ("group_id" by default), 
indicating the name of the colData(x) element which stores the group id of each cell (i.e., colData(x)$name_colData_group).}

\item{logarithm}{a boolean, indicating whether to use raw counts (if FALSE, by default) or log(counts + 1) (if TRUE).}

\item{P}{the number of permutations to use.}

\item{N_breaks}{the number of breaks at which to evaluate the comulative density function.}

\item{min_non_zero_cells}{the minimum number of non-zero cells in each cluster for a gene to be evaluated.}
}
\value{
A \code{\linkS4class{data.frame}} object.
Columns `gene` and `cluster_id` contain the gene and cell-cluster name, while `p_val`, `p_adj.loc` and `p_adj.glb` report the raw p-values, locally and globally adjusted p-values, via Benjamini and Hochberg (BH) correction.
In locally adjusted p-values (`p_adj.loc`) BH correction is applied in each cluster separately, while in globally adjusted p-values (`p_adj.glb`) BH correction is performed to the results from all clusters.
}
\description{
\code{distinct_test} tests for differential state between two groups of samples.
}
\examples{
data("Kang_subset", package = "distinct")
Kang_subset

set.seed(61217)
res = distinct_test(
  x = Kang_subset, 
  name_assays_expression = "counts",
  name_cluster = "cell",
  name_group = "stim",
  name_sample = "sample_id",
  P = 10^3, 
  min_non_zero_cells = 20)

# Visualize results:
head(res)

}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
