% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_results.R
\name{top_results}
\alias{top_results}
\title{Test for differential state between two groups of samples, based on scRNA-seq data.}
\usage{
top_results(res, cluster = "all", significance = 0.01, global = TRUE)
}
\arguments{
\item{res}{a \code{\linkS4class{data.frame}} with results as returned from \code{\link{distinct_test}}.}

\item{cluster}{a character indicating the cluster(s) whose results have to be returned. 
Results from all clusters are returned by default ("all").}

\item{significance}{numeric, results with adjusted p-value < significance will be returned.}

\item{global}{logical indicating whether to filter results according to p_adj.glb (when TRUE), or p_adj.loc (when FALSE).}
}
\value{
A \code{\linkS4class{data.frame}} object.
Columns `gene` and `cluster_id` contain the gene and cell-cluster name, while `p_val`, `p_adj.loc` and `p_adj.glb` report the raw p-values, locally and globally adjusted p-values, via Benjamini and Hochberg (BH) correction.
In locally adjusted p-values (`p_adj.loc`) BH correction is applied in each cluster separately, while in globally adjusted p-values (`p_adj.glb`) BH correction is performed to the results from all clusters.
}
\description{
\code{top_results} returns the significant results obtained via \code{\link{distinct_test}}.
}
\examples{
# load the input data:
data("Kang_subset", package = "distinct")
Kang_subset

# create the design of the study:
samples = Kang_subset@metadata$experiment_info$sample_id
group = Kang_subset@metadata$experiment_info$stim
design = model.matrix(~group)
# rownames of the design must indicate sample ids:
rownames(design) = samples
design

# The group we would like to test for is in the second column of the design, therefore we will specify: column_to_test = 2

set.seed(61217)
res = distinct_test(
  x = Kang_subset, 
  name_assays_expression = "logcounts",
  name_cluster = "cell",
  design = design,
  column_to_test = 2,
  min_non_zero_cells = 20,
  n_cores = 2)

# Visualize significant results:
top_results(res)

# Visualize significant results from a specified cluster of cells:
top_results(res, cluster = "Dendritic cells")

}
\seealso{
\code{\link{plot_cdfs}}, \code{\link{plot_densities}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
